--- services/newsletter/app/__main__.py
@@ -1,7 +1,0 @@
-from report import create_report
-
-def main() -> None:
-    create_report()
-
-if __name__ == "__main__":
-    main()
